temp_bot/requirements.txt
twitchio
twitchio[starlette]
python-dotenv
requests
pytest
pytest-asyncio
